requirements.txt
asgiref==3.4.1
catboost==0.26.1
cycler==0.10.0
Django==3.2.7
graphviz==0.17
joblib==1.0.1
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
pickle-mixin==1.0.2
Pillow==8.3.2
plotly==5.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tenacity==8.0.1
threadpoolctl==2.2.0
